==> code/2_eda.R <==
library(dplyr)
library(tidyverse)
library(here)
library(lubridate)
library(forcats)
library(ggplot2)
library(survival)
library(climate)

# data <- read_csv(here::here("data", "dd_data.csv")) %>% 
#   mutate(
#     imb_id = as.factor(imb_id),
#     host_plant = as.factor(host_plant),
#     # overall_survival = as.factor(overall_survival),
#     rearing_year = as.integer(rearing_year),
#     stage = as.factor(stage)
#   ) %>% 
#   mutate_at(vars(contains("stage")), 
#             factor,
#             levels = c("first", "second", "third", "fourth", "fifth", "pupa", "adult", "death"),
#             ordered = TRUE) 

temp_data <- read_csv(here::here("data", "mean_temp_data.csv")) %>% 
  mutate(
    imb_id = as.factor(imb_id),
    host_plant = as.factor(host_plant),
    # overall_survival = as.factor(overall_survival),
    rearing_year = as.integer(rearing_year),
    stage = as.factor(stage)
  ) %>% 
  mutate_at(vars(contains("stage")), 
            factor,
            levels = c("first", "second", "third", "fourth", "fifth", "pupa", "adult", "death"),
            ordered = TRUE) 

# plot of mean temperatures for each stage
mean_temp_plot <- ggplot(data = temp_data, aes(x = mean_temp, fill = stage)) +
  geom_density(alpha = 0.4) +
  theme_bw()
mean_temp_plot

duration_temp_plot <- ggplot(temp_data %>% filter(stage != "pupa") %>% mutate(surv = if_else(stage2 == "death", "Death", "Survival to next stage")), aes(x = mean_temp)) +
  geom_density(aes(fill = surv), alpha = 0.5) +
  # geom_smooth(method = "lm") +
  theme_bw() +
  facet_wrap(~stage) +labs(x = "Mean Temperature", y = "Density", fill = "Outcome") +
  theme(legend.position = c(0.8, 0.1)) +
  scale_fill_manual(values = c("#D95B6A", "#5BDA77"))
duration_temp_plot

duration_temp_plot <- ggplot(temp_data %>% filter(stage == "pupa", rearing_year != 2024, mispupation == "N") %>% mutate(surv = if_else(stage2 == "death", "Death", "Survival to next stage")), aes(x = mean_temp)) +
  geom_density(aes(fill = surv), alpha = 0.5) +
  # geom_smooth(method = "lm") +
  theme_bw() +
  facet_wrap(~stage) +labs(x = "Mean Temperature", y = "Density", fill = "Outcome") +
  theme(legend.position = c(0.8, 0.6)) +
  scale_fill_manual(values = c("#D95B6A", "#5BDA77"))
duration_temp_plot

weird_data <- temp_data %>% 
  filter(stage == "pupa", 
         mean_temp > 11)

# plot of mean temperatures for each stage
mean_temp_plot <- ggplot(data = temp_data %>% filter(stage == "pupa"), aes(x = mean_temp, fill = )) +
  geom_density(alpha = 0.4) +
  theme_bw()
mean_temp_plot

# plot of degree days by year
# temps <- station_temps_cleaned %>% 
#   mutate(year = year(date),
#          month = month(date),
#          day0 = mdy(paste0("01/01/", year)),
#          dayofYear = as.integer(date - day0),
#          degree_days = if_else(degree_days == "NaN", 0, degree_days),
#          year = as.factor(year)) %>% 
#   filter(year != "2024") %>% 
#   group_by(year) %>% 
#   mutate(cum_dd = cumsum(degree_days))

# ggplot(data = temps, aes(x = dayofYear, y = cum_dd)) +
#       geom_line(color = "darkgreen", linewidth = 1.3) +
#   facet_wrap(~year) +
#   theme_bw() +
#   labs(x = "day of year (0 = Jan. 1st)", y = "cumulative degree days")


# dd_data %>% group_by(imb_id) %>% 
#   slice(1) %>% 
#   ungroup() %>% 
#   count(overall_survival)


# ggplot(data = data, aes(x = stage)) +
#   geom_bar(stat = "count", aes(fill = survival)) +
#   theme_bw()
# 
# ggplot(data = data %>% filter(stage != "pupa"), aes(x = total_degree_days)) +
#   geom_density(aes(fill = stage), alpha = 0.4) +
#   theme_bw()

ggplot(data = data %>% filter(stage != "pupa"), aes(x = duration)) +
  geom_density(aes(fill = stage), alpha = 0.4) +
  theme_bw()

# cor(dd_data$duration, dd_data$total_degree_days)

# ggplot(data = data %>% filter(stage != "pupa"), aes(x = duration, y = total_degree_days)) +
#   geom_point() +
#   geom_smooth(method = "glm")
# 
# ggplot(data = data %>% filter(stage == "pupa"), aes(x = duration, y = total_degree_days)) +
#   geom_point() +
#   geom_smooth(method = "glm")



# reading in temperature data from rearing rooms. deleting from data prep because we're using station temps
threshold <- 10 # really not sure what the temperature threshold is supposed to be...
# cannot find minimum development threshold for IMB, but i found a source that generally says the min threshold for lots
# of butterflies is 15

temp_data <- read_csv(here::here("data", "continuous_temps.csv")) %>% 
  mutate(location = `...1`,
         date = ymd(date),
         time = hms(time),
         temp = (temp.f - 32) * (5/9)) %>% 
  filter(!str_detect(location, "mystery")) %>% 
  select(date, time, temp) %>% 
  group_by(date) %>% 
  summarise(min = min(temp),
            max = max(temp),
            mean_temp = mean(temp)) %>% 
  ungroup() %>% 
  mutate(degree_days = pmax(0, ((max+min)/2) - threshold))

summary(temp_data$date)

ggplot(temp_data %>% filter(date > mdy("03272016")), aes(x = date)) +
  geom_density()
# testing correlations between rearing room and weather station data to see if extrapolation is reasonable
# requires getting weatehr station temps by running data prep first

temp_comparisons <- left_join(temp_data, station_temps_cleaned, 
                              by = join_by(date == date),
                              suffix = c("_rearing_room", "_weather_station")) %>% 
  select(-min, -max, -degree_days)

# corr <- glm(mean ~ mean_temp,
#             data = temp_comparisons)
# summary(corr)

c <- cor.test(temp_comparisons$mean_temp_rearing_room, temp_comparisons$mean_temp_weather_station,
         method = "pearson")
c

c <- lm(mean_temp_rearing_room ~ mean_temp_weather_station,
        data = temp_comparisons)
c

plot(c)

ggplot(data = temp_comparisons, aes(x = mean_temp_rearing_room, y = mean_temp_weather_station)) +
  geom_point(alpha = 0.7) +
  geom_smooth(method = "glm") +
  labs(x = "Rearing Room Temperatures (°C)", y = "Friday Harbor Airport Temperatures (°C)") +
  scale_x_continuous(limits = c(-10, 30)) +
  scale_y_continuous(limits = c(-10, 30)) +
  theme_bw(base_size = 14)
